move add document storage adapter method
