+++ document-storage-adapter/src/Pssg.DocumentStorageAdapter.Client/Pssg.DocumentStorageAdapter.Client.csproj
@@ -12,7 +12,9 @@
 	  <PackageReference Include="Grpc.Tools" Version="2.62.0">
 		<PrivateAssets>all</PrivateAssets>
 		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	</PackageReference>		  
+	</PackageReference>
+	  <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
+	  <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />		  
   </ItemGroup>
